server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
celery==5.4.0
redis==5.2.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
xai-sdk>=1.5.0
apscheduler==3.10.4
